// src/types/DAS/das.ts
export declare namespace DAS {
// ...
  type GetAssetRequest = {
    id: string;
    displayOptions?: GetAssetDisplayOptions;
  };
// ...
  type GetAssetResponse = {
    interface: Interface;
    id: string;
    content?: Content;
    authorities?: Authorities[];
    compression?: Compression;
    grouping?: Grouping[];
    royalty?: Royalty;
    ownership: Ownership;
    creators?: Creators[];
    uses?: Uses;
    supply?: Supply;
    mutable: boolean;
    burnt: boolean;
    mint_extensions?: MintExtensions;
    token_info?: TokenInfo;
  };
// ...
  type GetAssetDisplayOptions = {
    showUnverifiedCollections?: boolean;
    showCollectionMetadata?: boolean;
    showRawData?: boolean;
    showFungible?: boolean;
    requireFullIndex?: boolean;
    showSystemMetadata?: boolean;
    showNativeBalance?: boolean;
    showInscription?: boolean;
  };
  interface Ownership {
    frozen: boolean;
    delegated: boolean;
    delegate?: string;
    ownership_model: OwnershipModel;
    owner: string;
  }
  interface Supply {
    print_max_supply: number;
    print_current_supply: number;
    edition_nonce?: number;
  }
  interface Uses {
    use_method: UseMethods;
    remaining: number;
    total: number;
  }
  interface Creators {
    address: string;
    share: number;
    verified: boolean;
  }
  interface Royalty {
    royalty_model: RoyaltyModel;
    target?: string;
    percent: number;
    basis_points: number;
    primary_sale_happened: boolean;
    locked: boolean;
  }
  interface Grouping {
    group_key: string;
    group_value: string;
    verified?: boolean;
    collection_metadata?: CollectionMetadata;
  }
  interface CollectionMetadata {
    name?: string;
    symbol?: string;
    image?: string;
    description?: string;
    external_url?: string;
  }
  interface Authorities {
    address: string;
    scopes: Array<Scope>;
  }
  type Links = {
    external_url?: string;
    image?: string;
    animation_url?: string;
    [Symbol.iterator](): Iterator<Links>;
  };
  interface Content {
    $schema: string;
    json_uri: string;
    files?: Files;
    metadata: Metadata;
    links?: Links;
  }
// ...
  interface Compression {
    eligible: boolean;
    compressed: boolean;
    data_hash: string;
    creator_hash: string;
    asset_hash: string;
    tree: string;
    seq: number;
    leaf_id: number;
  }
// ...
  interface MintExtensions {
    confidential_transfer_mint?: ConfidentialTransferMint;
    confidential_transfer_fee_config?: ConfidentialTransferFeeConfig;
    transfer_fee_config?: TransferFeeConfig;
    metadata_pointer?: MetadataPointer;
    mint_close_authority?: MintCloseAuthority;
    permanent_delegate?: PermanentDelegate;
    transfer_hook?: TransferHook;
    interest_bearing_config?: InterestBearingConfig;
    default_account_state?: DefaultAccountState;
    confidential_transfer_account?: ConfidentialTransferAccount;
    metadata?: MintExtensionMetadata;
  }
// ...
  interface TokenInfo {
    symbol?: string;
    balance?: number;
    supply?: number;
    decimals?: number;
    token_program?: string;
    associated_token_address?: string;
    price_info?: PriceInfo;
    mint_authority?: string;
    freeze_authority?: string;
  }
  interface PriceInfo {
    price_per_token: number;
    currency: string;
    total_price?: number;
  }
}